finetuning/requirements.txt
numpy
torch
torchvision
pillow
tqdm
matplotlib
tdw
ftfy
regex
sentence_transformers
